Split history websocket API into its own file
