+++ homeassistant/components/history/const.py
@@ -1,0 +1,10 @@
+"""History integration constants."""
+
+DOMAIN = "history"
+HISTORY_FILTERS = "history_filters"
+HISTORY_ENTITIES_FILTER = "history_entities_filter"
+HISTORY_USE_INCLUDE_ORDER = "history_use_include_order"
+EVENT_COALESCE_TIME = 0.35
+
+CONF_ORDER = "use_include_order"
+MAX_PENDING_HISTORY_STATES = 2048
